Newidea

+++ src/main/java/mTrepka/libary/service/BookServiceImpl.java
@@ -1,21 +1,25 @@
 package mTrepka.libary.service;
 
+import lombok.RequiredArgsConstructor;
 import mTrepka.libary.domain.Book;
+import mTrepka.libary.domain.BorrowHistory;
 import mTrepka.libary.domain.User;
 import mTrepka.libary.repository.BookRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import mTrepka.libary.repository.BorrowHistoryRepository;
 import org.springframework.stereotype.Service;
 
-import java.sql.Connection;
+import java.time.LocalDate;
 import java.util.List;
 
 /**
  * Created by Mario on 2017-07-10.
  */
 @Service("bookService")
+@RequiredArgsConstructor
 public class BookServiceImpl implements BookService{
-    @Autowired
-    private BookRepository bookRepository;
+	private final BookRepository bookRepository;
+	private final BorrowHistoryRepository borrowHistoryRepository;
+    private final UserService userService;
 
     @Override
     public List<Book> getAllBooks() {
@@ -27,15 +31,7 @@ public class BookServiceImpl implements BookService{
         return bookRepository.findById(id);
     }
 
-    @Override
-    public List<Book> findByOwnerid(User ownerid) {
-        return bookRepository.findByOwnerid(ownerid);
-    }
 
-    @Override
-    public void saveBook(Book book) {
-        bookRepository.save(book);
-    }
 
     @Override
     public List<Book> findAllFreeBook() {
@@ -43,12 +39,60 @@ public class BookServiceImpl implements BookService{
     }
 
     @Override
-    public List<Book> findAllBorrowBook() {
-        return bookRepository.findBorrowBook();
+    public void addBook(Book book) {
+        bookRepository.save(book);
     }
 
     @Override
-    public void removeBook(Book book) {
-        bookRepository.delete(book);
+    public void borrowBook(Long bookId) {
+        Book book = bookRepository.findById(bookId);
+	    if (book.getCurrentBorrow() == null) {
+            User user = userService.getCurrentUser();
+            BorrowHistory borrowHistory = new BorrowHistory();
+            borrowHistory.setBorrow_date(LocalDate.now());
+		    borrowHistory.setBookId(book.getId());
+		    borrowHistory.setUserCardNumber(user.getCardNumber());
+		    borrowHistory.setStatus("lend");
+            book.setCurrentBorrow(borrowHistory);
+		    user.getBooks().add(book);
+		    borrowHistoryRepository.save(borrowHistory);
+        }
     }
+
+    @Override
+    public List<Book> findAllBorrowBook() {
+	    return bookRepository.findBorrowBook();
+    }
+
+    @Override
+    public String removeBook(Book book) {
+	    if (book.getCurrentBorrow() == null) {
+            bookRepository.delete(book);
+            return "Book has been removed";
+        } else {
+            return "You can't remove book becouse is borrowed";
+        }
+
+    }
+
+    @Override
+    public Book editBookAndSave(Book bookData, long bookId) {
+        Book currentBook = bookRepository.findById(bookId);
+        editBookFromTemplate(currentBook, bookData);
+	    bookRepository.save(currentBook);
+        return currentBook;
+    }
+
+    private void editBookFromTemplate(Book currentBook, Book bookData) {
+        currentBook.setAuthor(bookData.getAuthor());
+        currentBook.setName(bookData.getName());
+        currentBook.setIsbn(bookData.getIsbn());
+        currentBook.setDescription(bookData.getDescription());
+        currentBook.setPageAmount(bookData.getPageAmount());
+        currentBook.setReleaseDate(bookData.getReleaseDate());
+    }
+
+	public List<Book> getLastFiveBooks() {
+		return bookRepository.getLastFiveBooks();
+	}
 }
